Fix: 計算式の修正

+++ src/main/java/com/tiscon/dao/EstimateDao.java
@@ -128,7 +128,7 @@ public class EstimateDao {
      * @param boxNum 総段ボール数
      * @return 料金[円]
      */
-    public int getPricePerTruck(int boxNum) {
+    public double getPricePerTruck(int boxNum) {
         String sql = "SELECT PRICE FROM TRUCK_CAPACITY WHERE MAX_BOX >= :boxNum ORDER BY PRICE LIMIT 1";
 
         SqlParameterSource paramSource = new MapSqlParameterSource("boxNum", boxNum);
